+++ yazi-plugin/preset/plugins/clipboard.lua
@@ -33,4 +33,31 @@ function M.copy_uri_list(list)
 	end
 end
 
+-- TODO: remove
+local path = tostring(rt.path.config_dir:join("plugins/clipboard.yazi/main.lua"))
+local try_compat = ya.sync(function(state, path)
+	if state._compat_loaded then
+		return true
+	end
+
+	local ok = pcall(dofile, path)
+	if ok then
+		ya.notify {
+			title = "Heads up:",
+			content = "`clipboard` is deprecated as a custom plugin name in favor of a built-in one.\n\nRename your `plugins/clipboard.yazi` to a different name, and update your `keymap.toml` to use the new name instead.",
+			level = "warn",
+			timeout = 20,
+		}
+	end
+
+	state._compat_loaded = ok
+	return ok
+end)
+
+function M:entry(job)
+	if try_compat(path) then
+		return dofile(path):entry(job)
+	end
+end
+
 return M